Refactor XML component imports to streamline module structure and enhance clarity
